feat: rename all tasks/triggers and improve doc.
Fixes #58
