+++ src/main/java/io/kestra/plugin/redis/list/ListPopInterface.java
@@ -1,10 +1,8 @@
-package io.kestra.plugin.redis;
+package io.kestra.plugin.redis.list;
 
 import io.kestra.core.models.annotations.PluginProperty;
-import io.kestra.plugin.redis.models.SerdeType;
 import io.swagger.v3.oas.annotations.media.Schema;
 
-import jakarta.validation.constraints.NotNull;
 import java.time.Duration;
 
 public interface ListPopInterface extends ListPopBaseInterface {
@@ -12,16 +10,19 @@ public interface ListPopInterface extends ListPopBaseInterface {
         title = "The max number of rows to fetch before stopping.",
         description = "It's not an hard limit and is evaluated every second."
     )
+    @PluginProperty
     Integer getMaxRecords();
 
     @Schema(
         title = "The max duration waiting for new rows.",
         description = "It's not an hard limit and is evaluated every second."
     )
+    @PluginProperty
     Duration getMaxDuration();
 
     @Schema(
         title = "Number of elements that should be pop at once"
     )
+    @PluginProperty
     Integer getCount();
 }
