+++ src/main/java/io/kestra/plugin/redis/list/package-info.java
@@ -1,0 +1,7 @@
+@PluginSubGroup(
+    description = "This sub-group of plugins contains tasks for using the Redis NoSQL database lists commands.",
+    categories = PluginSubGroup.PluginCategory.DATABASE
+)
+package io.kestra.plugin.redis.list;
+
+import io.kestra.core.models.annotations.PluginSubGroup;
