--- src/main/java/io/kestra/plugin/redis/package-info.java
@@ -1,7 +1,0 @@
-@PluginSubGroup(
-    description = "This sub-group of plugins contains tasks for using the Redis NoSQL database.",
-    categories = PluginSubGroup.PluginCategory.DATABASE
-)
-package io.kestra.plugin.redis;
-
-import io.kestra.core.models.annotations.PluginSubGroup;
