+++ src/main/java/io/kestra/plugin/redis/pubsub/package-info.java
@@ -1,0 +1,7 @@
+@PluginSubGroup(
+    description = "This sub-group of plugins contains tasks for using the Redis NoSQL database pub/sub commands.",
+    categories = PluginSubGroup.PluginCategory.DATABASE
+)
+package io.kestra.plugin.redis.pubsub;
+
+import io.kestra.core.models.annotations.PluginSubGroup;
