+++ src/main/java/io/kestra/plugin/redis/string/Delete.java
@@ -1,4 +1,4 @@
-package io.kestra.plugin.redis;
+package io.kestra.plugin.redis.string;
 
 import io.kestra.core.models.annotations.Example;
 import io.kestra.core.models.annotations.Plugin;
@@ -6,6 +6,7 @@ import io.kestra.core.models.annotations.PluginProperty;
 import io.kestra.core.models.executions.metrics.Counter;
 import io.kestra.core.models.tasks.RunnableTask;
 import io.kestra.core.runners.RunContext;
+import io.kestra.plugin.redis.AbstractRedisConnection;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
@@ -31,7 +32,8 @@ import java.util.List;
                 "   - keyDelete2"
             }
         )
-    }
+    },
+    aliases = "io.kestra.plugin.redis.Delete"
 )
 public class Delete extends AbstractRedisConnection implements RunnableTask<Delete.Output> {
     @Schema(
@@ -45,6 +47,7 @@ public class Delete extends AbstractRedisConnection implements RunnableTask<Dele
         title = "If some keys are not deleted, failed the task"
     )
     @Builder.Default
+    @PluginProperty
     private Boolean failedOnMissing = false;
 
     @Override
