+++ src/main/java/io/kestra/plugin/redis/string/package-info.java
@@ -1,0 +1,7 @@
+@PluginSubGroup(
+    description = "This sub-group of plugins contains tasks for using the Redis NoSQL database strings commands.",
+    categories = PluginSubGroup.PluginCategory.DATABASE
+)
+package io.kestra.plugin.redis.string;
+
+import io.kestra.core.models.annotations.PluginSubGroup;
