+++ src/test/java/io/kestra/plugin/redis/DeleteTest.java
@@ -3,6 +3,8 @@ package io.kestra.plugin.redis;
 import com.google.common.collect.ImmutableMap;
 import io.kestra.core.runners.RunContext;
 import io.kestra.core.runners.RunContextFactory;
+import io.kestra.plugin.redis.string.Delete;
+import io.kestra.plugin.redis.string.Set;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
 import jakarta.inject.Inject;
 import org.junit.jupiter.api.Assertions;
