+++ src/test/java/io/kestra/plugin/redis/ListPopTest.java
@@ -4,7 +4,10 @@ import com.google.common.collect.ImmutableMap;
 import io.kestra.core.runners.RunContext;
 import io.kestra.core.runners.RunContextFactory;
 import io.kestra.core.storages.StorageInterface;
+import io.kestra.plugin.redis.list.ListPop;
+import io.kestra.plugin.redis.list.ListPush;
 import io.kestra.plugin.redis.models.SerdeType;
+import io.kestra.plugin.redis.string.Delete;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
 import jakarta.inject.Inject;
 import org.junit.jupiter.api.BeforeEach;
