+++ src/test/resources/flows/realtime.yaml
@@ -3,7 +3,7 @@ namespace: io.kestra.tests
 
 triggers:
   - id: watch
-    type: io.kestra.plugin.redis.RealtimeTriggerList
+    type: io.kestra.plugin.redis.list.RealtimeTrigger
     url: redis://:redis@localhost:6379/0
     key: mytriggerkey
